+++ src-opam/dockerfile_distro.mli
@@ -31,8 +31,27 @@ type t = [
   | `OracleLinux of [ `V7 | `V8 | `Latest ]
   | `OpenSUSE of [ `V42_1 | `V42_2 | `V42_3 | `V15_0 | `V15_1 | `V15_2 | `Latest ]
   | `Ubuntu of [ `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_04 | `V16_10 | `V17_04 | `V17_10 | `V18_04 | `V18_10 | `V19_04 | `V19_10 | `V20_04 | `V20_10 | `LTS | `Latest ]
+  | `Windows of [ `V20H2 | `Latest ]
 ] [@@deriving sexp]
 (** Supported Docker container distributions *)
+
+type os_family = [ `Linux | `Windows ] [@@deriving sexp]
+(** Supported Docker container operating systems  *)
+
+val os_family_of_distro : t -> os_family
+(** [os_family_of_distro t] returns the OS family of the distro. *)
+
+val os_family_to_string : os_family -> string
+(** [os_family_to_string os] returns a string representing the OS
+   family. *)
+
+val opam_repository : os_family -> string
+(** [opam_repository os_family] returns the git URL to the default
+   Opam repository. *)
+
+val personality : os_family -> Ocaml_version.arch -> string option
+(** [personality os_family arch] returns the personality associated to
+   the architecture, if [os_family] is [`Linux]. *)
 
 val compare : t -> t -> int
 (** [compare a b] is a lexical comparison function for {!t}. *)
@@ -70,6 +89,7 @@ type package_manager = [
   | `Yum  (** Fedora Yum *)
   | `Zypper (** OpenSUSE Zypper *)
   | `Pacman (** Archlinux Pacman *)
+  | `Cygwin (** Cygwin package manager *)
 ] [@@deriving sexp]
 (** The package manager used by a distro. *)